// src/models/id.ts
export type Identification = {
    fullName: string
    firstContact: {
        name: string
        phone: string
    }
    secondContact: {
        name: string
        phone: string
    }
    address: string
    diagnosticLink: string
    additionalInfo: string
    cid: string
    dateOfBirth: string
}